resume-backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.23.2
pydantic==2.11.7
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
motor==3.3.2
pymongo==4.6.1
pdfplumber==0.9.0
python-docx==0.8.11
httpx==0.24.1
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==23.2.1
apscheduler==3.10.4
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.8.3
pillow==10.4.0
groq==0.11.0
